\name{qep-package}
\alias{qep-package}
\alias{qep}
\docType{package}
\title{
\packageTitle{qep}
}
\description{
\packageDescription{qep}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{qep}
\packageIndices{qep}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{qep}

Maintainer: \packageMaintainer{qep}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
